fix: Invalid key error for metric_filter and dimension_filter (#187)
This PR fixes the issue introduced by #186.

It was causing the tap to break if no metric filter or dimensions are
defined in the reports.json, or if these keys are defined with
_camelCase_ notation, which would be reasonable given it's a JSON file
and all the other dimension and metric values use _camelCase_ notation.

The issue was that the keys were defined with _camelCase_ on lines
113-114 on the `client.py` file when initializing the
`report_definition` dictionary, but it then tries to access them using
_snake_case_ notation on lines 267-268, which caused the tap to break
with an invalid key error.